requirements.txt
streamlit>=1.30
torch>=2.0
torchvision>=0.15
pandas>=1.5
numpy>=1.23
matplotlib>=3.7
Pillow>=9.5
requests>=2.31
